Delete records

+++ priv/repo/migrations/20240315094657_alter_waste.exs
@@ -5,8 +5,8 @@ defmodule EcoService.Repo.Migrations.AlterWaste do
     alter table(:wastes) do
       remove :sef_lf_bags, :integer
       add :seg_lf_bags, :integer
-      remove :kg_of_sef_lf, :integer
-      add :kg_of_seg_lf, :integer
+      remove :kg_of_sef_lf, :decimal
+      add :kg_of_seg_lf, :decimal
     end
   end
 end
